api-main/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
attrs==21.4.0
beautifulsoup4==4.11.1
blis==0.7.7
bs4==0.0.1
cachetools==5.0.0
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
coverage==6.3.2
cssselect==1.1.0
cymem==2.0.6
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.6.0
Flask==2.1.1
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.5
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.1
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
langcodes==3.3.0
libclang==13.0.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
murmurhash==1.0.6
newspaper3k==0.2.8
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pathy==0.6.1
Pillow==9.1.0
pluggy==1.0.0
preshed==3.0.6
protobuf==3.20.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==3.0.8
pytest==7.1.1
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.0.15
tinysegmenter==0.3
tldextract==3.2.1
tomli==2.0.1
tqdm==4.64.0
typer==0.4.1
typing_extensions==4.1.1
urllib3==1.26.9
wasabi==0.9.1
Werkzeug==2.1.1
wrapt==1.14.0
